Worked to incorporate the semantics of call-by-need into the intermediate representation, and generally improved correctness of the IR translator/evaluator. *Nearly* all tests pass with the IR-based evaluator, but not quite so it is still switched off.
I also fixed some issues with the complex number tests. complex() was being used
incorectly in the tests, the first parameter is actually the length of the output vector,
not the real component. I'm not sure why the tests were passing before-- i think i must 
have fixed one of my bugs that was obscuring the failure. Anyway, I've corrected the
tests and verified that it produces results consistent with C-R.





git-svn-id: http://renjin.googlecode.com/svn/trunk@523 b722d886-ae41-2056-5ac6-6db9963f4c99